#18 - 좋아요 싫어요 기능 상속으로 분리
