+++ src/main/java/com/kgat/entity/Article.java
@@ -16,7 +16,7 @@ import java.util.Objects;
 @AllArgsConstructor
 @Entity
 @Table(name="articles")
-public class Article {
+public class Article extends ReactionContent {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long articleNum;
@@ -39,19 +39,7 @@ public class Article {
     @Column(nullable = false)
     private int viewCount;
 
-    @Column(nullable = false)
-    private int likeCount;
 
-    @Column(nullable = false)
-    private int hateCount;
-
-    @Transient
-    @JsonProperty("isLike")
-    private boolean isLike = false;
-
-    @Transient
-    @JsonProperty("isHate")
-    private boolean isHate = false;
 
     @PrePersist
     protected void onCreate() {
@@ -74,10 +62,7 @@ public class Article {
                 ", createdAt=" + createdAt +
                 ", updatedAt=" + updatedAt +
                 ", viewCount=" + viewCount +
-                ", likeCount=" + likeCount +
-                ", hateCount=" + hateCount +
-                ", isLike=" + isLike +
-                ", isHate=" + isHate +
+
                 '}';
     }
 
@@ -99,48 +84,5 @@ public class Article {
     public int hashCode() {
         return Objects.hash(articleNum);
     }
-
-    public void toggleLike() {
-        if(isLike) {
-            // 이미 좋아요를 누른 상태에서 클릭하면
-            // 좋아요 비활성화, 카운트 감소
-            likeCount--;
-            isLike = false;
-        } else {
-            // 좋아요를 누르지 않았던 상태에서 클릭하면
-            // 좋아요 활성화, 카운트 증가
-            likeCount++;
-            isLike = true;
-
-            // 싫어요 활성화 상태면
-            // 비활성화, 카운트 감소
-            if(isHate) {
-                hateCount--;
-                isHate = false;
-            }
-        }
-    }
-
-    public void toggleHate() {
-        if(isHate) {
-            // 이미 싫어요를 누른 상태에서 클릭하면
-            // 싫어요 비활성화, 카운트 감소
-            hateCount--;
-            isHate = false;
-        } else {
-            // 싫어요를 누르지 않은 상태에서 클릭하면
-            // 싫어요 활성화, 카운트 증가
-            hateCount++;
-            isHate = true;
-        }
-        
-        if(isLike) {
-            // 좋아요가 활성화 된 상태라면
-            // 좋아요 비활성화, 좋아요 카운트 감소
-            likeCount--;
-            isLike = false;
-        }
-    }
-
 
 }
