+++ src/main/java/com/kgat/entity/Toggleable.java
@@ -1,0 +1,6 @@
+package com.kgat.entity;
+
+public interface Toggleable {
+    void toggleLike();
+    void toggleHate();
+}
